% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion.R
\name{accordion}
\alias{accordion}
\title{Accordion Script}
\usage{
accordion(input_id)
}
\arguments{
\item{input_id}{(\code{character(1)})\cr
string with which to register a shiny input. The title of the currently open panel (or
last opened panel) will be accessible with \code{input$input_id}.}
}
\description{
Javascript to make a panel group into an accordion where
only one panel item can be open at once.
}
\keyword{internal}
