% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_plot_modifications}
\alias{apply_plot_modifications}
\title{This function checks the plot type and applies specific modifications
to the plot object based on the provided parameters.}
\usage{
apply_plot_modifications(plot_obj, plot_type, dblclicking, ranges)
}
\arguments{
\item{plot_obj}{The original plot object.}

\item{plot_type}{The type of the plot, either \code{gg} (\code{ggplot2}) or \code{grob} (\code{grid}, \code{graphics}).}

\item{dblclicking}{A logical value indicating whether double-clicking on data points on
the main plot is enabled or disabled.}

\item{ranges}{A list containing x and y values of ranges.}
}
\description{
This function checks the plot type and applies specific modifications
to the plot object based on the provided parameters.
}
\keyword{internal}
