% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_table_args.R
\name{basic_table_args}
\alias{basic_table_args}
\title{Builds a \code{basic_table_args} object}
\usage{
basic_table_args(...)
}
\arguments{
\item{...}{arguments compatible with \code{\link[rtables:basic_table]{rtables::basic_table()}}.}
}
\value{
(\code{basic_table_args}) object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function has to be used to build an input for a \code{basic_table_args} argument.
The \code{basic_table_args} argument should be a part of every module which contains any \code{rtables} object.
Arguments are validated to match their \code{rtables} equivalents.

For more details see the vignette: \code{vignette("custom-basic-table-arguments", package = "teal.widgets")}.
}
\examples{
basic_table_args(subtitles = "SUBTITLE")
}
\seealso{
\itemize{
\item \code{\link[=resolve_basic_table_args]{resolve_basic_table_args()}} to resolve multiple objects into one using pre-defined priorities.
\item \code{\link[=parse_basic_table_args]{parse_basic_table_args()}} to parse resolved list into list of calls.
}
}
