% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_with_settings.R
\name{clean_brushedPoints}
\alias{clean_brushedPoints}
\title{Cleans and organizes output to account for NAs and remove empty rows.}
\usage{
clean_brushedPoints(data, brush)
}
\arguments{
\item{data}{(\code{data.frame})\cr
A dataframe from which to select rows.}

\item{brush}{(\code{list})\cr
The data from a brush e.g. input$plot_brush.}
}
\value{
A dataframe of selected rows.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
