% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbatim_popup.R
\name{disable_button}
\alias{disable_button}
\title{Disables a button}
\usage{
disable_button(disabled_flag, button_id)
}
\arguments{
\item{disabled_flag}{(\code{reactive}) containing the flag}

\item{button_id}{(\code{character(1)}) the id of the controlled button}
}
\description{
Disables a button
}
\details{
When the flag is \code{TRUE} the button to open the popup is disabled; it is enabled otherwise.
}
\keyword{internal}
