% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbatim_popup.R
\name{disabled_flag_observer}
\alias{disabled_flag_observer}
\title{Creates a \code{shiny} observer handling the disabled flag.}
\usage{
disabled_flag_observer(disabled_flag, button_id, ignoreInit = TRUE)
}
\arguments{
\item{disabled_flag}{(\code{reactive}) containing the flag}

\item{button_id}{(\code{character(1)}) the id of the controlled button}

\item{ignoreInit}{(\code{logical(1)}) should the observer be ignored on initialization}
}
\description{
Creates a \code{shiny} observer handling the disabled flag.
}
\details{
When the flag is \code{TRUE} the button to open the popup is disabled; it is enabled otherwise.
}
\keyword{internal}
