% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draggable_buckets.R
\name{draggable_buckets}
\alias{draggable_buckets}
\title{Draggable Buckets}
\usage{
draggable_buckets(input_id, elements, buckets)
}
\arguments{
\item{input_id}{(\code{character(1)}) the \code{HTML} id of this widget}

\item{elements}{(\code{character}) the elements to drag into buckets}

\item{buckets}{(\code{character}) the names of the buckets the elements can but put in}
}
\value{
the \code{HTML} code comprising an instance of this widget
}
\description{
A custom widget with draggable elements that can be put into buckets.
}
\examples{
ui <- shiny::fluidPage(
  draggable_buckets("id", c("a", "b"), c("bucket1", "bucket2")),
  # draggable_buckets("id2", c("a", "b"), c("bucket1", "bucket2")),
  shiny::verbatimTextOutput("out")
)
server <- function(input, output) {
  shiny::observeEvent(input$id, {
    print(shiny::isolate(input$id))
  })
  output$out <- shiny::renderPrint(input$id)
}
if (interactive()) shiny::shinyApp(ui, server)

}
