% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionalInput.R
\name{extract_raw_choices}
\alias{extract_raw_choices}
\title{Extract raw values from choices}
\usage{
extract_raw_choices(choices, sep)
}
\arguments{
\item{choices}{(\code{choices_labeled}, \code{list} or \code{character})\cr
object containing choices}

\item{sep}{(\code{character(1)})\cr
A separator string to split the \code{choices} or \code{selected} inputs into the values of
the different columns.}
}
\value{
choices simplified
}
\description{
Extract raw values from choices
}
\keyword{internal}
