% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_name_validation_srv}
\alias{file_name_validation_srv}
\title{Validate the download file name}
\usage{
file_name_validation_srv(input, output)
}
\arguments{
\item{input, output}{Shiny module \code{input} and \code{output} objects.}
}
\value{
\code{reactive} returning \code{TRUE} when the file name is valid.
}
\description{
Hides the \code{data_download} button and renders \code{output$file_name_warning} when \code{input$file_name}
is not meaningful, i.e. it has 3 characters or fewer or contains only special characters or whitespace.
}
\keyword{internal}
