% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbatim_popup.R
\name{format_content}
\alias{format_content}
\title{Formats the content of the modal popup window.}
\usage{
format_content(verbatim_content)
}
\arguments{
\item{verbatim_content}{(\code{character}, \code{expression} or \code{reactive(1)} holding either)
the content to show in the popup modal window}
}
\value{
\code{reactive} with the formatted content
}
\description{
Formats the content of the modal popup window.
}
\details{
Formats the content:
\itemize{
\item concatenates if needed
\item styles
}
}
\keyword{internal}
