% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dt_rows.R
\name{get_dt_rows}
\alias{get_dt_rows}
\title{Maps the \code{lengthMenu}selected value property of \code{DT::datatable} to a Shiny variable.}
\usage{
get_dt_rows(dt_name, dt_rows)
}
\arguments{
\item{dt_name}{\code{ns()} of inputId of the \code{DT::datatable}}

\item{dt_rows}{\code{ns()} of inputId of the variable that holds the current selected value of \code{lengthMenu}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
library(shiny)
# in Shiny UI function
tagList(
  get_dt_rows(ns("data_table"), ns("dt_rows")),
  ...
)

# use the input$dt_rows in the Shiny Server function
if (!is.null(input$dt_rows)) {
  dt_args$options$pageLength <- input$dt_rows
} # nolint
do.call(DT::datatable, dt_args)
}

}
