% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionalInput.R
\name{optionalSelectInput}
\alias{optionalSelectInput}
\title{Wrapper on \code{pickerInput}}
\usage{
optionalSelectInput(
  inputId,
  label = NULL,
  choices = NULL,
  selected = NULL,
  multiple = FALSE,
  sep = NULL,
  options = list(),
  label_help = NULL,
  fixed = FALSE,
  width = NULL
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{choices}{(\code{character}, \code{NULL})\cr
If \code{choices} length is less or equal to 1, the \code{pickerInput} widget is hidden and
non-interactive widget will be displayed instead.}

\item{selected}{The initially selected value (or multiple values if \code{multiple = TRUE}).
If not specified then defaults to the first value for single-select lists
and no values for multiple select lists.}

\item{multiple}{Is selection of multiple items allowed?}

\item{sep}{(\code{character(1)})\cr
A separator string to split the \code{choices} or \code{selected} inputs into the values of the different
columns.}

\item{options}{List of options, see \link[shinyWidgets]{pickerOptions} for all available options.
To limit the number of selection possible, see example below.}

\item{label_help}{(\code{shiny.tag} optional)\cr
e.g. an object returned by \code{\link[shiny:helpText]{shiny::helpText()}}.}

\item{fixed}{(\code{logical(1)} optional)\cr
whether to block user to select choices.}

\item{width}{(\code{character(1)})\cr
The width of the input passed to \code{pickerInput}  e.g. \code{'auto'}, \code{'fit'}, \code{'100px'} or \code{'75\%'}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Wrapper on \code{\link[shinyWidgets:pickerInput]{shinyWidgets::pickerInput()}} with additional features.
When \code{fixed = TRUE} or when the number of \code{choices} is less or equal to 1,
the \code{pickerInput} widget is hidden and non-interactive widget will be displayed
instead. With change of the user interface \code{selected} remains unchanged.
Toggle of \code{HTML} elements is just the visual effect to avoid displaying
\code{pickerInput} widget when there is only one choice.
}
\examples{
library(shiny)

# Create a minimal example data frame
data <- data.frame(
  AGE = c(25, 30, 40, 35, 28),
  SEX = c("Male", "Female", "Male", "Female", "Male"),
  PARAMCD = c("Val1", "Val2", "Val3", "Val4", "Val5"),
  PARAM = c("Param1", "Param2", "Param3", "Param4", "Param5"),
  AVISIT = c("Visit1", "Visit2", "Visit3", "Visit4", "Visit5"),
  stringsAsFactors = TRUE
)

app <- shinyApp(
  ui = fluidPage(
    optionalSelectInput(
      inputId = "c1",
      label = "Default",
      choices = LETTERS[1:5],
      selected = "A"
    ),
    optionalSelectInput(
      inputId = "c2",
      label = "Fixed choices",
      choices = LETTERS[1:5],
      selected = "A",
      fixed = TRUE
    ),
    optionalSelectInput(inputId = "c3", label = "Single choice", choices = "A", selected = "A"),
    optionalSelectInput(inputId = "c4", label = "NULL choices", choices = NULL),
    optionalSelectInput(
      inputId = "c5",
      label = "Named vector",
      choices = c(`A - value A` = "A", `B - value B` = "B", `C - value C` = "C"),
      selected = "A"
    ),
    optionalSelectInput(
      inputId = "c6",
      label = "Variable choices",
      choices = teal.transform::variable_choices(data = data, subset = c("AGE", "SEX", "PARAMCD")),
      selected = "PARAMCD"
    ),
    optionalSelectInput(
      inputId = "c7",
      label = "Value choices",
      choices = teal.transform::value_choices(
        data = data,
        var_choices = c("PARAMCD", "AVISIT"),
        var_label = c("PARAM", "AVISIT")
      ),
      selected = "Val1 - Visit1"
    ),
    verbatimTextOutput(outputId = "out")
  ),
  server = function(input, output, session) {
    output$out <- renderText({
      paste0(
        "Default: ", input$c1, "\n",
        "Fixed choices: ", input$c2, "\n",
        "Single choice: ", input$c3, "\n",
        "NULL choices: ", input$c4, "\n",
        "Named vector: ", input$c5, "\n",
        "Variable choices: ", input$c6, "\n",
        "Value choices: ", input$c7, "\n"
      )
    })
  }
)

if (interactive()) {
  runApp(app)
}
}
