% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_group.R
\name{panel_status}
\alias{panel_status}
\title{Get Collapsible Panel Status}
\usage{
panel_status(input_id, div_id, panel_id)
}
\arguments{
\item{input_id}{(\code{character})\cr
string with which to register a shiny input. The title of the currently open panel (or last
opened panel) will be accessible with \code{input$input_id}.}

\item{div_id}{(\code{character})\cr
string giving the id of the \code{div} to listen for collapse events on.}

\item{panel_id}{(\code{character})\cr
string giving the id of the panel to determine the initial collapsed status.}
}
\description{
Javascript to get open or close status of a panel item.
}
\keyword{internal}
