% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_table_args.R
\name{parse_basic_table_args}
\alias{parse_basic_table_args}
\title{Parses \code{basic_table_args} object into the \code{basic_table} expression}
\usage{
parse_basic_table_args(basic_table_args = teal.widgets::basic_table_args())
}
\arguments{
\item{basic_table_args}{(\code{basic_table_args})\cr
This argument could be a result of the \code{\link[=resolve_basic_table_args]{resolve_basic_table_args()}}.}
}
\value{
(\code{language}) the \code{rtables::basic_table()} filled with additional arguments.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
A function to parse expression from the \code{basic_table_args} object.
}
\examples{
parse_basic_table_args(
  resolve_basic_table_args(
    user_table = basic_table_args(title = "TITLE"),
    user_default = basic_table_args(title = "DEFAULT_TITLE", subtitles = "SUBTITLE")
  )
)
}
