% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionalInput.R
\name{picker_options_content}
\alias{picker_options_content}
\title{Optional content for \code{optionalSelectInput}}
\usage{
picker_options_content(var_name, var_label, var_type)
}
\arguments{
\item{var_name}{(\code{character})\cr
variable name}

\item{var_label}{(\code{character})\cr
variable alternative name - for example variable label}

\item{var_type}{(\code{character})
class of the variable.}
}
\value{
(\code{character}) HTML contents with all elements combined
}
\description{
Prepares content to be displayed in \code{optionalSelectInput} with icons and labels
}
\keyword{internal}
