% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_suppress}
\alias{plot_suppress}
\title{This function opens a PDF graphics device using \code{\link[grDevices]{pdf}} to suppress
the plot display in the IDE. The purpose of this function is to avoid opening graphic devices
directly in the IDE.}
\usage{
plot_suppress(x)
}
\arguments{
\item{x}{lazy binding which generates the plot(s)}
}
\description{
This function opens a PDF graphics device using \code{\link[grDevices]{pdf}} to suppress
the plot display in the IDE. The purpose of this function is to avoid opening graphic devices
directly in the IDE.
}
\keyword{internal}
