% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_with_settings.R
\name{plot_with_settings_ui}
\alias{plot_with_settings_ui}
\title{UI part of plot-with-settings module}
\usage{
plot_with_settings_ui(id)
}
\arguments{
\item{id}{(\code{character}) module ID}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
