% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_with_settings.R
\name{print_plot}
\alias{print_plot}
\title{Print plot for download functionality}
\usage{
print_plot(plot, plot_type)
}
\arguments{
\item{plot}{(\code{reactive})\cr
reactive expression to draw a plot}

\item{plot_type}{(\code{reactive})\cr
reactive plot type (\code{gg}, \code{trel}, \code{grob}, \code{other})}
}
\value{
Nothing returned, the plot is printed.
}
\description{
Print plot for download functionality
}
\keyword{internal}
