% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_with_settings.R
\name{table_with_settings_ui}
\alias{table_with_settings_ui}
\title{UI part of table-with-settings module}
\usage{
table_with_settings_ui(id, ...)
}
\arguments{
\item{...}{(\code{character})\cr
Useful for providing additional HTML classes for the output tag.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
