% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white_small_well.R
\name{white_small_well}
\alias{white_small_well}
\title{Small well class for HTML}
\usage{
white_small_well(...)
}
\arguments{
\item{...}{other arguments to pass to tag object's div attributes.}
}
\value{
An HTML output element with class Small Well and overflow-x property
}
\description{
Adds Small Well class and overflow-x property to HTML output element.
}
\details{
\code{white_small_well} is intended to be used with \code{\link[shiny:uiOutput]{shiny::uiOutput()}}.
The overflow-x property is set to auto so that a scroll bar is added
when the content overflows at the left and right edges of the output window.
For example, this is useful for displaying wide tables.
}
\examples{

white_small_well(shiny::htmlOutput("summary"))
}
