% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbatim_popup.R
\name{button_click_observer}
\alias{button_click_observer}
\title{Creates a \code{shiny} observer handling button clicks.}
\usage{
button_click_observer(
  click_event,
  copy_button_id,
  copied_area_id,
  modal_title,
  modal_content,
  disabled
)
}
\arguments{
\item{click_event}{\code{reactive} the click event}

\item{copy_button_id}{(\code{character(1)}) the id of the button to copy the modal content.
Automatically appended with a 1 and 2 suffix for top and bottom buttons respectively.}

\item{copied_area_id}{(\code{character(1)}) the id of the element which contents are copied}

\item{modal_title}{(\code{character(1)}) the title of the modal window}

\item{modal_content}{(\code{reactive}) the content of the modal window}

\item{disabled}{(\code{reactive(1)}) the \code{shiny} reactive value holding a \code{logical}. The popup button is disabled
when the flag is \code{TRUE} and enabled otherwise.}
}
\description{
When the button is clicked it pop up a modal window with the text.
}
\keyword{internal}
