% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbatim_popup.R
\name{format_content}
\alias{format_content}
\title{Formats the content of the modal popup window.}
\usage{
format_content(verbatim_content, style = FALSE)
}
\arguments{
\item{verbatim_content}{(\code{character}, \code{expression}, \code{condition} or \code{reactive(1)}
holding any of the above) the content to show in the popup modal window}

\item{style}{(\code{logical(1)}) whether to style the \code{verbatim_content} using \code{styler::style_text}.
If \code{verbatim_content} is a \code{condition} or \code{reactive} holding \code{condition} then this argument is ignored}
}
\value{
\code{reactive} with the formatted content
}
\description{
Formats the content of the modal popup window.
}
\details{
Formats the content:
\itemize{
\item concatenates if needed
\item styles if \code{style} is TRUE
}
}
\keyword{internal}
