% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_group.R
\name{panel_group}
\alias{panel_group}
\title{Panel group widget}
\usage{
panel_group(..., id = NULL)
}
\arguments{
\item{...}{(\code{shiny.tag})\cr
panels created by \code{\link[=panel_group]{panel_group()}}}

\item{id}{optional, (\code{character})\cr}
}
\value{
(\code{shiny.tag})
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
