% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_group.R
\name{panel_group}
\alias{panel_group}
\title{Panel group widget}
\usage{
panel_group(..., id = NULL)
}
\arguments{
\item{...}{(\code{shiny.tag})\cr
panels created by \code{\link[=panel_group]{panel_group()}}}

\item{id}{optional, (\code{character})\cr}
}
\value{
(\code{shiny.tag})
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}\cr
Designed to group \code{\link{panel_item}} elements. Used to handle \code{shiny} inputs in the encoding panel.
}
\examples{

library(shiny)
panel_group(
  panel_item(
    title = "Display",
    collapsed = FALSE,
    checkboxGroupInput(
      "check",
      "Tables display",
      choices = LETTERS[1:3],
      selected = LETTERS[1]
    ),
    radioButtons(
      "radio",
      label = "Plot type",
      choices = letters[1:2],
      selected = letters[1]
    )
  ),
  panel_item(
    title = "Pre-processing",
    radioButtons(
      "filtering",
      "What to filter",
      choices = LETTERS[1:4],
      selected = LETTERS[1]
    ),
    radioButtons(
      "na_action",
      "NA action",
      choices = letters[1:3],
      selected = letters[1]
    )
  )
)

}
