% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.widgets-package.R
\docType{package}
\name{teal.widgets-package}
\alias{teal.widgets}
\alias{teal.widgets-package}
\title{\code{teal.widgets}: Custom widgets for \code{teal} applications}
\description{
The \code{teal.widgets} package contains functionalities that can be used
for \code{teal} module development. The package is dedicated to allow developers to extend module
functionality by additional inputs and output controls.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://insightsengineering.github.io/teal.widgets/}
  \item \url{https://github.com/insightsengineering/teal.widgets}
  \item Report bugs at \url{https://github.com/insightsengineering/teal.widgets/issues}
}

}
\author{
\strong{Maintainer}: Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}

Authors:
\itemize{
  \item Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}
  \item Pawel Rucki \email{pawel.rucki@roche.com}
  \item Mahmoud Hallal \email{mahmoud.hallal@roche.com}
  \item Nikolas Burkoff
  \item Maciej Nasinski
  \item Konrad Pagacz
  \item Junlue Zhao
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
