% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.widgets-package.R
\docType{package}
\name{teal.widgets-package}
\alias{teal.widgets}
\alias{teal.widgets-package}
\title{teal.widgets: Custom widgets for teal applications}
\description{
The \code{teal.widgets} package contains functionalities that can be used
for teal module development. The Package is dedicated to allow developers to extend modules
functionality by additional inputs and output controls.
}
\author{
\strong{Maintainer}: NEST \email{basel.nestcicd@roche.com}

Authors:
\itemize{
  \item Nikolas Burkoff \email{nikolas.burkoff@roche.com}
  \item Mahmoud Hallal \email{mahmoud.hallal@roche.com}
  \item Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}
  \item Maciej Nasinski \email{maciej.nasinski@contractors.roche.com}
  \item Konrad Pagacz \email{konrad.pagacz@contractors.roche.com}
  \item Pawel Rucki \email{pawel.rucki@contractors.roche.com}
  \item Junlue Zhao \email{zhaoj88@gene.com}
}

Other contributors:
\itemize{
  \item Tadeusz Lewandowski \email{tadeusz.lewandowski@roche.com} [contributor]
  \item Roche [copyright holder, funder]
}

}
\keyword{internal}
